# Use ReadFull over CopyN when decoding GGUFs