Opting to use `io.ReadFull` with preallocated `[]byte` slices as `bytes.Buffer` requires multiple grows/allocs to read in long strings. 

In addition, preallocating the array slices to prevent underlying allocs to append to the slice. 

I observed real world performance improvements as well as by using a small microbenchmark reading phi3 4Q from disk (file was cached by the OS) I see the following improvements:
```
goos: linux
goarch: amd64
pkg: github.com/ollama/ollama/llm
cpu: 12th Gen Intel(R) Core(TM) i7-1260P
          │   old.out   │               new.out               │
          │   sec/op    │   sec/op     vs base                │
Decode-16   63.57m ± 2%   35.28m ± 2%  -44.50% (p=0.000 n=10)

          │    old.out    │               new.out                │
          │     B/op      │     B/op      vs base                │
Decode-16   58.691Mi ± 0%   3.195Mi ± 0%  -94.56% (p=0.000 n=10)

          │   old.out   │               new.out               │
          │  allocs/op  │  allocs/op   vs base                │
Decode-16   324.4k ± 0%   227.1k ± 0%  -29.97% (p=0.000 n=10)
```
Results my vary depending on disk performance and whether the file was cached by the OS.

I may dig around for future performance improvements in the future.