@brycereitano 2024-04-26
Updated to load the data arrays in a more efficient manner. I don't think the array values are ever used and we could likely seek over them but I didn't want to include that in this PR. Here are the update microbenchmark results:
```
goos: linux
goarch: amd64
pkg: github.com/ollama/ollama/llm
cpu: 12th Gen Intel(R) Core(TM) i7-1260P
          │   old.out   │               mid.out               │               new.out               │
          │   sec/op    │   sec/op     vs base                │   sec/op     vs base                │
Decode-16   63.41m ± 2%   36.54m ± 1%  -42.37% (p=0.000 n=10)   21.43m ± 2%  -66.21% (p=0.000 n=10)

          │    old.out    │               mid.out                │               new.out                │
          │     B/op      │     B/op      vs base                │     B/op      vs base                │
Decode-16   58.690Mi ± 0%   3.195Mi ± 0%  -94.56% (p=0.000 n=10)   3.891Mi ± 0%  -93.37% (p=0.000 n=10)

          │   old.out   │               mid.out               │               new.out               │
          │  allocs/op  │  allocs/op   vs base                │  allocs/op   vs base                │
Decode-16   324.4k ± 0%   227.1k ± 0%  -29.97% (p=0.000 n=10)   131.0k ± 0%  -59.60% (p=0.000 n=10)
```

In ideal conditions it reliably loads data >2x faster which allow Ollama to determine fit faster bringing down the time to first character on larger models significantly.